:sparkles: Add ListContainer to split presentational component
