+++ src/App.jsx
@@ -5,13 +5,11 @@ import Page from './Page';
 import {
   updateTaskTitle,
   addTask,
-  deleteTask,
 } from './actions';
 
 function selector(state) {
   return {
     taskTitle: state.taskTitle,
-    tasks: state.tasks,
   };
 }
 
@@ -26,19 +24,13 @@ export default function App() {
     dispatch(addTask());
   }
 
-  function handleClickDeleteTask(id) {
-    dispatch(deleteTask(id));
-  }
-
-  const { taskTitle, tasks } = useSelector(selector);
+  const { taskTitle } = useSelector(selector);
 
   return (
     <Page
       taskTitle={taskTitle}
-      tasks={tasks}
       onChangeTitle={handleChangeTitle}
       onClickAddTask={handleClickAddTask}
-      onClickDeleteTask={handleClickDeleteTask}
     />
   );
 }
