+++ src/ListContainer.jsx
@@ -1,0 +1,26 @@
+import { useDispatch, useSelector } from 'react-redux';
+
+import List from './List';
+
+import {
+  deleteTask,
+} from './actions';
+
+export default function ListContainer() {
+  const { tasks } = useSelector((state) => ({
+    tasks: state.tasks,
+  }));
+
+  const dispatch = useDispatch();
+
+  function handleClickDeleteTask(id) {
+    dispatch(deleteTask(id));
+  }
+
+  return (
+    <List
+      tasks={tasks}
+      onClickDeleteTask={handleClickDeleteTask}
+    />
+  );
+}
