+++ src/Page.test.jsx
@@ -1,24 +1,27 @@
 import { render, fireEvent } from '@testing-library/react';
 
+import { useSelector } from 'react-redux';
+
 import Page from './Page';
 
+jest.mock('react-redux');
+
 test('Page', () => {
+  useSelector.mockImplementation((selector) => selector({
+    tasks: [
+      { id: 1, title: 'Task-1' },
+      { id: 2, title: 'Task-2' },
+    ],
+  }));
+
   const handleChangeTitle = jest.fn();
   const handleClickAddTask = jest.fn();
-  const handleClickDeleteTask = jest.fn();
-
-  const tasks = [
-    { id: 1, title: 'Task-1' },
-    { id: 2, title: 'Task-2' },
-  ];
 
   const { getByText } = render((
     <Page
       taskTitle=""
       onChangeTitle={handleChangeTitle}
       onClickAddTask={handleClickAddTask}
-      tasks={tasks}
-      onClickDeleteTask={handleClickDeleteTask}
     />
   ));
